[quant][graphmode] Pass debug option into insert_quant_dequant pass
Summary:
Some of the usage, e.g. add_scalar will not be supporting the debug option,
that is, we will not have a numerically exact representation of the final quantized model
before finalize if people use add scalar.
warning will be added in a later PR.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 214b60efd76e4d0c87ab1b3e73f0425b3ef152cc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39915