Eliminate TensorIterator::add_output with explicit dtype.
I'm working on a refactor where I want to represent the inputs
and outputs to TensorIterator as just plain Tensors, which means
I need to kill this add_output with explicit dtype.  This exists
solely to set what the output dtype should be.  We have a pretty
similar API for doing this for shapes (declare_static_shape) so
I just copied this API for dtypes instead.

Although the new version is more code, I think the intent is more
explicit.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: e57455d993dc6d231199917478a21ba5247f7f05
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39800